+++ lib/src/collect.dart
@@ -30,6 +30,9 @@ const _retryInterval = Duration(milliseconds: 200);
 /// If [includeDart] is true, code coverage for core `dart:*` libraries will be
 /// collected.
 ///
+/// If [functionCoverage] is true, function coverage information will be
+/// collected.
+///
 /// If [scopedOutput] is non-empty, coverage will be restricted so that only
 /// scripts that start with any of the provided paths are considered.
 ///
@@ -37,7 +40,9 @@ const _retryInterval = Duration(milliseconds: 200);
 /// those VM isolates.
 Future<Map<String, dynamic>> collect(Uri serviceUri, bool resume,
     bool waitPaused, bool includeDart, Set<String>? scopedOutput,
-    {Set<String>? isolateIds, Duration? timeout}) async {
+    {Set<String>? isolateIds,
+    Duration? timeout,
+    bool functionCoverage = false}) async {
   scopedOutput ??= <String>{};
 
   // Create websocket URI. Handle any trailing slashes.
@@ -72,7 +77,7 @@ Future<Map<String, dynamic>> collect(Uri serviceUri, bool resume,
     }
 
     return await _getAllCoverage(
-        service, includeDart, scopedOutput, isolateIds);
+        service, includeDart, functionCoverage, scopedOutput, isolateIds);
   } finally {
     if (resume) {
       await _resumeIsolates(service);
@@ -86,6 +91,7 @@ Future<Map<String, dynamic>> collect(Uri serviceUri, bool resume,
 Future<Map<String, dynamic>> _getAllCoverage(
     VmService service,
     bool includeDart,
+    bool functionCoverage,
     Set<String>? scopedOutput,
     Set<String>? isolateIds) async {
   scopedOutput ??= <String>{};
@@ -106,7 +112,7 @@ Future<Map<String, dynamic>> _getAllCoverage(
             isolateRef.id!, <String>[SourceReportKind.kCoverage],
             forceCompile: true, scriptId: script.id);
         final coverage = await _getCoverageJson(
-            service, isolateRef, scriptReport, includeDart);
+            service, isolateRef, scriptReport, includeDart, functionCoverage);
         allCoverage.addAll(coverage);
       }
     } else {
@@ -116,7 +122,7 @@ Future<Map<String, dynamic>> _getAllCoverage(
         forceCompile: true,
       );
       final coverage = await _getCoverageJson(
-          service, isolateRef, isolateReport, includeDart);
+          service, isolateRef, isolateReport, includeDart, functionCoverage);
       allCoverage.addAll(coverage);
     }
   }
@@ -204,13 +210,17 @@ Future<void> _processFunction(VmService service, IsolateRef isolateRef,
       print('tokenPos $tokenPos has no line mapping for script ${script.uri!}');
       return;
     }
-    hits.funcNames[line] = funcName;
+    hits.funcNames![line] = funcName;
   }
 }
 
 /// Returns a JSON coverage list backward-compatible with pre-1.16.0 SDKs.
-Future<List<Map<String, dynamic>>> _getCoverageJson(VmService service,
-    IsolateRef isolateRef, SourceReport report, bool includeDart) async {
+Future<List<Map<String, dynamic>>> _getCoverageJson(
+    VmService service,
+    IsolateRef isolateRef,
+    SourceReport report,
+    bool includeDart,
+    bool functionCoverage) async {
   final hitMaps = <Uri, HitMap>{};
   final scripts = <ScriptRef, Script>{};
   final libraries = <LibraryRef>{};
@@ -236,7 +246,13 @@ Future<List<Map<String, dynamic>>> _getCoverageJson(VmService service,
 
     // If the script's library isn't loaded, load it then look up all its funcs.
     final libRef = script.library;
-    if (libRef != null && !libraries.contains(libRef)) {
+    if (functionCoverage && libRef != null && !libraries.contains(libRef)) {
+      if (hits.funcHits == null) {
+        hits.funcHits = <int, int>{};
+      }
+      if (hits.funcNames == null) {
+        hits.funcNames = <int, String>{};
+      }
       libraries.add(libRef);
       final library =
           await service.getObject(isolateRef.id!, libRef.id!) as Library;
@@ -271,8 +287,8 @@ Future<List<Map<String, dynamic>>> _getCoverageJson(VmService service,
         continue;
       }
       _incrementCountForKey(hits.lineHits, line);
-      if (hits.funcNames.containsKey(line)) {
-        _incrementCountForKey(hits.funcHits, line);
+      if (hits.funcNames != null && hits.funcNames!.containsKey(line)) {
+        _incrementCountForKey(hits.funcHits!, line);
       }
     }
     for (final tokenPos in coverage.misses!) {
@@ -283,8 +299,8 @@ Future<List<Map<String, dynamic>>> _getCoverageJson(VmService service,
       }
       hits.lineHits.putIfAbsent(line, () => 0);
     }
-    hits.funcNames.forEach((line, funcName) {
-      hits.funcHits.putIfAbsent(line, () => 0);
+    hits.funcNames?.forEach((line, funcName) {
+      hits.funcHits?.putIfAbsent(line, () => 0);
     });
   }