+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Support function level coverage information, when running tests in the Dart
    VM. This is not supported for web tests yet.
+* Add flag `--function-coverage` (abbr `-f`) to collect_coverage that collects
+  function coverage information.
 * Add flag `--pretty-print-func` (abbr `-f`) to format_coverage that works
   similarly to pretty print, but outputs function level coverage, rather than
   line level.
@@ -10,7 +12,8 @@
 * BREAKING CHANGE: The signatures of `createHitmap`, `mergeHitmaps`,
   `parseCoverage`, `toScriptCoverageJson`, and `Formatter.format` have changed
   from using `Map<int, int>` to represent line coverage to using `HitMap`
-  (which contains both line and function coverage).
+  (which contains both line and function coverage). `collect` also has a new
+  optional bool flag controlling whether function coverage is collected.
 
 ## 1.0.4 - 2021-06-08